requirements.txt
pillow>=10.0.0
pypdf>=6.0.0
pikepdf>=9.0.0